see_matched_adoptables

+++ lib/match/mutation.py
@@ -1,5 +1,6 @@
 import graphene
 
+from bson import ObjectId
 from .type import MatchType
 from .model import Match
 
@@ -9,14 +10,16 @@ class CreateMatch(graphene.Mutation):
     match = graphene.Field(MatchType)
 
     class Arguments:
-        repository_id = graphene.String(required=True)
+        adoptable = graphene.String(required=True)
 
     def mutate(root, info, **args):
         token = getattr(info.context, "token", None)
-        match = Match(**args, user=token["username"])
-        match.save()
-
-        return CreateMatch(match=match)
+        if token is not None:
+            ObjectId(args.get("adoptable"))
+            match = Match(**args, user=token["username"])
+            match.save()
+            return CreateMatch(match=match)
+        return {}
 
 
 class DeleteMatch(graphene.Mutation):
@@ -24,14 +27,16 @@ class DeleteMatch(graphene.Mutation):
     match = graphene.Field(MatchType)
 
     class Arguments:
-        repository_id = graphene.String(required=True)
+        adoptable = graphene.String(required=True)
 
     def mutate(root, info, **args):
         token = getattr(info.context, "token", None)
-        match = Match.objects.get(**args, user=token["username"])
-        match.delete()
-
-        return DeleteMatch(match=match)
+        if token is not None:
+            ObjectId(args.get("adoptable"))
+            match = Match.objects.get(**args, user=token["username"])
+            match.delete()
+            return DeleteMatch(match=match)
+        return {}
 
 
 class Mutation:
